feat: New selector and source format for parsers (#304)
This PR introduces two new things:

The parser source can now supply an array that is concatenated using path.join()
The parser has a new syntax for querying a file in the selector property
The selector property is based on JMESPath and uses a specific language to query Javascript objects. This should make it easier to query more complexly nested data, as wel as be more performant in these complex cases.